delete with confirmation

+++ jweb-cms/page-admin/src/main/web/admin/page/template.list.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import {Link} from "react-router-dom";
-import {Button, Form, Input, Message as notification, Pagination, Table} from "element-react";
+import {Button, Form, Input, Message as notification, MessageBox, Pagination, Table} from "element-react";
+
 
 const i18n = window.i18n;
 export default class TemplateList extends React.Component {
@@ -111,37 +112,47 @@ export default class TemplateList extends React.Component {
 
     delete(data, e) {
         e.preventDefault();
-        fetch("/admin/api/page/" + data.id, {method: "DELETE"}).then(() => {
-            this.find();
-            notification({
-                title: i18n.t("page.successTitle"),
-                type: "success",
-                message: i18n.t("page.deleteSuccessMessage")
+
+        MessageBox.confirm(i18n.t("page.deletePageTip"), i18n.t("page.deleteHint"), {
+            type: 'warning'
+        }).then(() => {
+            fetch("/admin/api/page/" + data.id, {method: "DELETE"}).then(() => {
+                this.find();
+                notification({
+                    title: i18n.t("page.successTitle"),
+                    type: "success",
+                    message: i18n.t("page.deleteSuccessMessage")
+                });
             });
         });
     }
 
     batchDelete(e) {
         e.preventDefault();
-        const list = this.state.selected;
-        if (list.length === 0) {
-            return;
-        }
-        const ids = [];
-        for (let i = 0; i < list.length; i += 1) {
-            ids.push(list[i].id);
-        }
-        fetch("/admin/api/page/batch-delete", {
-            method: "POST",
-            body: JSON.stringify({ids: ids})
+
+        MessageBox.confirm(i18n.t("page.deletePageTip"), i18n.t("page.deleteHint"), {
+            type: 'warning'
         }).then(() => {
-            notification({
-                title: i18n.t("page.successTitle"),
-                type: "success",
-                message: i18n.t("page.deleteSuccessMessage")
+            const list = this.state.selected;
+            if (list.length === 0) {
+                return;
+            }
+            const ids = [];
+            for (let i = 0; i < list.length; i += 1) {
+                ids.push(list[i].id);
+            }
+            fetch("/admin/api/page/batch-delete", {
+                method: "POST",
+                body: JSON.stringify({ids: ids})
+            }).then(() => {
+                notification({
+                    title: i18n.t("page.successTitle"),
+                    type: "success",
+                    message: i18n.t("page.deleteSuccessMessage")
+                });
+                this.setState({selected: []});
+                this.find();
             });
-            this.setState({selected: []});
-            this.find();
         });
     }
